Reword the explanation for `orders` parameter
Currently, the `orders` parameter is a bit difficult to understand. I had to read the section multiple times to get it.
Can you help go through the edit to see if I understood it correctly, myself? If so, I would like to contribute to making it easier for others too.